platform/web_server/requirements.txt
alt-profanity-check~=1.2
amplitude-analytics~=0.3
boto3~=1.21
celery~=5.2
click~=8.0
Django~=3.2
django-bootstrap-v5~=1.0
django-celery-beat~=2.2
django-celery-results~=2.2
django-cors-headers~=3.13
django-extensions~=3.1
django-oauth-toolkit~=2.1
django-polymorphic~=3.1
django-rest-polymorphic~=0.1
django-steamauth>=1.1.2
django-storages~=1.12
djangorestframework~=3.13
gunicorn~=20.1
ipython~=7.31
piecemaker~=0.4
PSNAWP~=1.3
psd_tools~=1.9
psycopg2-binary~=2.9
pysha3~=1.0
pytest-bdd~=7.1
pytest-cov~=3.0
pytest-django~=4.5
pytest-factoryboy~=2.1
pytest-lazy-fixtures~=1.0
pytest-mock~=3.7
python-dotenv~=0.16
redis~=4.1
sentry-sdk~=1.30
uvicorn~=0.16
wagtail~=2.16
web3~=5.28
